+++ Changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `parity-scale-codec` support.
+
 ## [1.4.1] — 2022-10-15
 
 ### Changed